HAL LL ADC: Fix u32 => uptr pointer casts for 64-bit build

+++ shared/STM32MP2xx_HAL_Driver/Inc/stm32mp2xx_ll_adc.h
@@ -447,7 +447,7 @@ extern "C" {
   * @retval Pointer to register address
   */
 #define __ADC_PTR_REG_OFFSET(__REG__, __REG_OFFFSET__)                         \
-  ((__IO uint32_t *)((uint32_t) ((uint32_t)(&(__REG__)) + ((__REG_OFFFSET__) << 2UL))))
+  ((__IO uint32_t *)((uintptr_t) ((uintptr_t)(&(__REG__)) + ((__REG_OFFFSET__) << 2UL))))
 
 /**
   * @}
@@ -2127,12 +2127,12 @@ __STATIC_INLINE uint32_t LL_ADC_DMA_GetRegAddr(const ADC_TypeDef *ADCx, uint32_t
   if (Register == LL_ADC_DMA_REG_REGULAR_DATA)
   {
     /* Retrieve address of register DR */
-    data_reg_addr = (uint32_t)(&(ADCx->DR));
+    data_reg_addr = (uint32_t)(uintptr_t)(&(ADCx->DR));
   }
   else /* (Register == LL_ADC_DMA_REG_REGULAR_MULTI_NO_PACKING || LL_ADC_DMA_REG_REGULAR_MULTI_PACKING) */
   {
     /* Retrieve address of register CDR */
-    data_reg_addr = (uint32_t)(&((__LL_ADC_COMMON_INSTANCE(ADCx))->CDR));
+    data_reg_addr = (uint32_t)(uintptr_t)(&((__LL_ADC_COMMON_INSTANCE(ADCx))->CDR));
   }
 
   return data_reg_addr;
@@ -2144,7 +2144,7 @@ __STATIC_INLINE uint32_t LL_ADC_DMA_GetRegAddr(const ADC_TypeDef *ADCx, uint32_t
   (void)(Register);
 
   /* Retrieve address of register DR */
-  return (uint32_t)(&(ADCx->DR));
+  return (uint32_t)(uintptr_t)(&(ADCx->DR));
 }
 #endif /* ADC_MULTIMODE_SUPPORT */
